Use single font settings file and set font to terminus
Instead of having a standard and custom font file use a single xfont
file and update the script to write out the default font size to this
file if a file size is not passed in. Also, after a recent update to
arch it seems terminus12 is no longer working so change the font face to
terminus instead.